# TypedElementOnSourceNode: Caching of Children and Value to improve performance