|                     Method |     Mean |    Error |   StdDev |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------:|---------:|---------:|-----------:|----------:|------:|----------:|
| ToTypedElementOnSourceNode (before caching)| 885.8 ms | 26.63 ms | 77.26 ms | 23000.0000 | 3000.0000 |     - |    143 MB |
| ToTypedElementOnSourceNode (after caching)| 724.1 ms | 14.34 ms | 20.56 ms | 19000.0000 | 9000.0000 |     - |    117 MB |